+++ examples/futures/Trading/GetTradeFeeRate.php
@@ -13,7 +13,5 @@ $APIContract = new APIContractTrading(new CloudConfig([
     'memo' => "<your_memo>",
 ]));
 
-$response = $APIContract->cancelOrder("BTCUSDT", "<order_id>")['response'];
-
-
+$response = $APIContract->getContractTradeFeeRate('BTCUSDT')['response'];
 echo json_encode($response);