+++ tests/BitMart/Websocket/Spot/WsSpotPrvTest.php
@@ -7,28 +7,26 @@ use BitMart\Websocket\Spot\WsSpotPrv;
 
 include_once __DIR__ . '/../../../../vendor/autoload.php';
 
+
 // Cli Run: php tests/BitMart/Websocket/Spot/WsSpotPrvTest.php start
 $cloudPrivateWebsocket = new WsSpotPrv(
    [
        'accessKey' => "your_api_key",
        'secretKey' => "your_secret_key",
        'memo' =>  "your_memo",
-       'xdebug' => true
+       'xdebug' => true,
+       'callback' => function ($data) {
+           echo "-------------------------" . PHP_EOL;
+           print_r($data);
+       },
+       'pong' => function ($data) {
+           echo "-------------------------".$data.PHP_EOL;
+       }
    ]
 );
 
+// Login
+$cloudPrivateWebsocket->login();
 
 // Subscribe Private Channels
-$cloudPrivateWebsocket->subscribe(
-    [
-        'op' => "subscribe",
-        'args' => [
-            // Only Support Private Channel
-            "spot/user/order:BTC_USDT",
-        ]
-    ]       ,
-    function ($data) {
-        echo "-------------------------" . PHP_EOL;
-        print_r($data);
-    }
-);
+$cloudPrivateWebsocket->send('{"op": "subscribe", "args": ["spot/user/balance:BALANCE_UPDATE"]}');